211021 - #1

+++ endpoint/board/detail/model.js
@@ -4,13 +4,14 @@ class BoardDetailModel extends BoardModel {
     super(req);
 
     this.boardId = req.params?.boardId;
+    this.addHashtags = req.body?.addHashtags ?? [];
+    this.deleteHashtags = req.body?.deleteHashtags ?? [];
   }
 
   async delete(res) {
-    this.isAuthorized();
-
     this.checkParameters(this.boardId);
     await this.dao.serialize(async db => {
+      await this.checkAuthorized(db);
       const result = await db.run('delete from board where board.id=? and board.userId=?', [
         this.boardId, this.requestUserID
       ]);
@@ -28,10 +29,10 @@ class BoardDetailModel extends BoardModel {
       const boards = await db.get('select user.name, board.createdAt, board.content from board left join user on board.userId=user.id where board.id=?', [
         this.boardId
       ]);
-      if(!boards.length) {
+      const board = boards[0];
+      if(!board) {
         throw new Error('404 내용 없음 예정');
       }
-      const board = boards[0];
 
       const hashtags = await db.get('select boardHashtag.hashtag from boardHashtag where boardHashtag.boardId=?', [
         this.boardId
@@ -60,43 +61,42 @@ class BoardDetailModel extends BoardModel {
   }
 
   async update(res) {
-    /*
-    const userId = req.user?.id;
-    if(userId===undefined) {
-      throw new Error('401 Unauthorized (예정)');
-    }
-    const boardId = req.params?.boardId;
-    const content = req.body?.content;
-    const addHashtags = req.body?.addHashtags ?? [];
-    const deleteHashtags = req.body?.deleteHashtags ?? [];
-    if(content===undefined || boardId===undefined) {
-      throw new Error('400 파라미터 오류(예정)');
-    }
-    const query = maria('query');
-    query('update board set board.content=? where board.id=? and board.userId=?', [
-      content, boardId, userId
-    ])(result => {
+    this.checkParameters(this.boardId, this.content);
+    await this.dao.serialize(async db => {
+      await this.checkAuthorized(db);
+      const result = await db.run('update board set board.content=? where board.id=? and board.userId=?', [
+        this.content, this.boardId, this.requestUserID
+      ]);
       if(!result.affectedRows) {
         throw new Error('403 권한 없음 예정');
       }
-    })
-    for(const hashtag of addHashtags) {
-      query('insert into boardHashtag(boardId, hashtag) values (?, ?)', [
-        boardId, hashtag
-      ]);
-    }
-    for(const hashtag of deleteHashtags) {
-      query('delete from boardHashtag where boardHashtag.boardId=? and boardHashtag.hashtag=?', [
-        boardId, hashtag
-      ]);
-    }
-    query(() => {
+
+      const tags = (await db.get('select boardHashtag.hashtag from boardHashtag where boardHashtag.boardId=?', [
+        this.boardId
+      ])).map(row => row.hashtag);
+
+      for(const hashtag of this.addHashtags) {
+        console.log(hashtag, tags, tags.includes(hashtag));
+        if(tags.includes(hashtag)) {
+          continue;
+        }
+        await db.run('insert into boardHashtag(boardId, hashtag) values (?, ?)', [
+          this.boardId, hashtag
+        ]);
+      }
+      for(const hashtag of this.deleteHashtags) {
+        if(!tags.includes(hashtag)) {
+          continue;
+        }
+        await db.run('delete from boardHashtag where boardHashtag.boardId=? and boardHashtag.hashtag=?', [
+          this.boardId, hashtag
+        ]);
+      }
+
       res.json({
-        boardId,
         complete: true
       });
-    })().catch(err => next(err));
-    */
+    });
   }
 }
 module.exports = BoardDetailModel;
